+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Extensions/RazorCodeDocumentExtensions.cs
@@ -3,7 +3,9 @@
 
 using System;
 using System.Diagnostics;
+using System.Linq;
 using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.AspNetCore.Razor.Language.Intermediate;
 using Microsoft.CodeAnalysis.Razor.Protocol;
 using Microsoft.CodeAnalysis.Text;
 
@@ -130,5 +132,15 @@ internal static class RazorCodeDocumentExtensions
 
         csharpRange = default;
         return false;
+    }
+
+    public static bool ComponentNamespaceMatches(this RazorCodeDocument razorCodeDocument, string fullyQualifiedNamespace)
+    {
+        var namespaceNode = (NamespaceDeclarationIntermediateNode)razorCodeDocument
+            .GetDocumentIntermediateNode()
+            .FindDescendantNodes<IntermediateNode>()
+            .First(n => n is NamespaceDeclarationIntermediateNode);
+
+        return namespaceNode.Content == fullyQualifiedNamespace;
     }
 }
